refactor: unify PageHeader component location
Move PageHeader from `~/entrypoints/options/components/PageHeader` to
`~/components/PageHeader` and update all imports across the codebase to
use the centralized component location.
