+++ entrypoints/options/pages/AccountManagement/index.tsx
@@ -2,8 +2,8 @@ import { CalendarCheck2, UserRound } from "lucide-react"
 import { type MouseEvent } from "react"
 import { useTranslation } from "react-i18next"
 
+import { PageHeader } from "~/components/PageHeader"
 import { Button } from "~/components/ui"
-import { PageHeader } from "~/entrypoints/options/components/PageHeader"
 import AccountList from "~/features/AccountManagement/components/AccountList"
 import { useAccountActionsContext } from "~/features/AccountManagement/hooks/AccountActionsContext"
 import { useAccountDataContext } from "~/features/AccountManagement/hooks/AccountDataContext"
