+++ features/About/About.tsx
@@ -9,10 +9,10 @@ import { useTranslation } from "react-i18next"
 
 import FeatureList from "~/components/FeatureList"
 import LinkCard from "~/components/LinkCard"
+import { PageHeader } from "~/components/PageHeader"
 import { Heading4 } from "~/components/ui"
 import { FEATURES, FUTURE_FEATURES } from "~/constants/about"
 import { EXTENSION_STORE_LISTING_URLS } from "~/constants/extensionStores"
-import { PageHeader } from "~/entrypoints/options/components/PageHeader"
 import packageJson from "~/package.json"
 import { isNotEmptyArray } from "~/utils"
 import type { ExtensionStoreId } from "~/utils/browser"
